+++ source/source_cell/test/klist_test.cpp
@@ -12,8 +12,8 @@
 #include "source_cell/setup_nonlocal.h"
 #include "source_cell/unitcell.h"
 #include "source_estate/magnetism.h"
-#include "source_pw/module_pwdft/VL_in_pw.h"
-#include "source_pw/module_pwdft/VNL_in_pw.h"
+#include "source_pw/module_pwdft/vl_pw.h"
+#include "source_pw/module_pwdft/vnl_pw.h"
 #include "source_pw/module_pwdft/parallel_grid.h"
 #include "source_io/berryphase.h"
 #include "source_io/module_parameter/parameter.h"
