+++ source/source_esolver/esolver_fp.h
@@ -10,7 +10,7 @@
 #include "source_estate/elecstate.h" // electronic states
 #include "source_estate/module_charge/charge_extra.h" // charge extrapolation
 #include "source_hamilt/module_surchem/surchem.h" // solvation model
-#include "source_pw/module_pwdft/VL_in_pw.h" // local pseudopotential
+#include "source_pw/module_pwdft/vl_pw.h" // local pseudopotential
 #include "source_pw/module_pwdft/structure_factor.h" // structure factor
 
 #include <fstream>
