+++ source/source_esolver/esolver_ks_pw.h
@@ -2,7 +2,7 @@
 #define ESOLVER_KS_PW_H
 #include "./esolver_ks.h"
 #include "source_psi/setup_psi_pw.h" // mohan add 20251012
-#include "source_pw/module_pwdft/VSep_in_pw.h"
+#include "source_pw/module_pwdft/vsep_pw.h"
 #include "source_pw/module_pwdft/exx_helper.h"
 #include "source_pw/module_pwdft/op_pw_vel.h"
