+++ source/source_estate/test/elecstate_pw_test.cpp
@@ -6,7 +6,7 @@
 #define protected public
 #include "source_estate/elecstate_pw.h"
 #include "source_hamilt/module_xc/xc_functional.h"
-#include "source_pw/module_pwdft/VL_in_pw.h"
+#include "source_pw/module_pwdft/vl_pw.h"
 #include "source_io/module_parameter/parameter.h"
 // mock functions for testing
 int XC_Functional::func_type = 1;
