+++ source/source_psi/psi_initializer.h
@@ -2,7 +2,7 @@
 #define PSI_INITIALIZER_H
 // data structure support
 #include "source_basis/module_pw/pw_basis_k.h" // for kpoint related data structure
-#include "source_pw/module_pwdft/VNL_in_pw.h"
+#include "source_pw/module_pwdft/vnl_pw.h"
 #include "source_pw/module_pwdft/structure_factor.h"
 #include "source_psi/psi.h" // for psi data structure
 // smart pointer for auto-memory management
