+++ source/source_pw/module_pwdft/op_pw_nl.h
@@ -7,7 +7,7 @@
 #include "source_pw/module_pwdft/kernels/nonlocal_op.h"
 #include "source_base/kernels/math_kernel_op.h"
 
-#include "source_pw/module_pwdft/VNL_in_pw.h"
+#include "source_pw/module_pwdft/vnl_pw.h"
 
 namespace hamilt {
