+++ source/source_pw/module_pwdft/stress_func.h
@@ -11,7 +11,7 @@
 #include "source_basis/module_pw/pw_basis_k.h"
 #include "source_cell/klist.h"
 #include "source_estate/module_charge/charge.h"
-#include "source_pw/module_pwdft/VNL_in_pw.h"
+#include "source_pw/module_pwdft/vnl_pw.h"
 #include "source_pw/module_pwdft/kernels/stress_op.h"
 #include "source_pw/module_pwdft/structure_factor.h"
 #include "source_base/kernels/math_kernel_op.h"
