+++ source/source_pw/module_pwdft/vsep_pw.cpp
@@ -1,4 +1,4 @@
-#include "VSep_in_pw.h"
+#include "vsep_pw.h"
 
 #include "source_base/constants.h"
 #include "source_base/libm/libm.h"
