+++ source/source_pw/module_stodft/sto_stress_pw.h
@@ -3,7 +3,7 @@
 
 #include "source_basis/module_pw/pw_basis_k.h"
 #include "source_estate/elecstate.h"
-#include "source_pw/module_pwdft/VL_in_pw.h"
+#include "source_pw/module_pwdft/vl_pw.h"
 #include "source_pw/module_pwdft/stress_func.h"
 #include "sto_wf.h"
